Adding dark mode support.
This initial version is done based on the OS theme (so it'll only be dark mode if the OS setting is dark mode).

In order for this to work I've added the customised less files for semantic locally and this requires a change
from the react build scripts to craco because react doesn't support less compilation.

I've verified as many screens as I can to make sure they're all looking ok but I'm sure I missed some parts.

CodeMirror needs a slightly different solution which needed JS code to detect the current theme.

The color scheme is heavily based on githubs dark mode and is using mostly css variable support. This should
make it easy to tweak in the future too.